Move expr tests to exprs.rs
Now that we have a separate module to contain the logic to compile
expressions, move the relevant tests into it even if the tests are
not exactly unit tests.
